move transaction writer out of insert_historical_blocK
